Extract CssDependency location into a trait
Use that trait in the LessPHP filter as well.
