+++ src/Filter/LessPHP.php
@@ -1,9 +1,9 @@
 <?php
 namespace AssetCompress\Filter;
 
-use lessc;
-
 use AssetCompress\AssetFilter;
+use AssetCompress\Filter\CssDependencyTrait;
+use lessc;
 
 /**
  * Pre-processing filter that adds support for LESS.css files.
@@ -14,9 +14,11 @@ use AssetCompress\AssetFilter;
  */
 class LessPHP extends AssetFilter
 {
+    use CssDependencyTrait;
 
     protected $_settings = array(
         'ext' => '.less',
+        'paths' => [],
     );
 
     /**
